Refresh page from server (Issue #6)
This is in response to @luc-ass / Issue #6. 

Adding line 273 refreshes the page (from the server, not from the cache) when the "Generate" button is pressed. This resets the form values, preventing web browsers from retaining cached information and producing duplicate profiles (even when the form input is changed by the user). 

I have tested creating numerous profiles with this change, and all have been produced as intended. After your testing, I believe this bug can be resolved.